requirements.txt
requests~=2.32.3
pyodbc~=5.2.0
bs4~=0.0.1
beautifulsoup4~=4.12.3
Unidecode~=1.3.8
futures~=3.0.5
PyPDF2~=3.0.1
python-dotenv~=1.0.0
openpyxl~=3.1.5